RAXAJSS-566
Changed the LabOrders UI so that it has plus and minus buttons that work like the Pharmacy ones.
Clicking the plus button adds not just a lab order textbox but a form which includes an Urgent checkbox.
Also removed the space in between the Lab Order textbox and the Urgent checkbox so that it is clear which order the urgency applies to.